// components/CTreeTransfer/interface.ts
export enum TransferCheckStrategy {
  all = 'all',
  parent = 'parent',
  child = 'child',
}

export enum LoadType {
  remoteExpand = 'remoteExpand',
}
// 节点数据
export interface TreeTransferDataType extends TreeNodeProps {
  title: string | React.ReactNode;
  key: string;
  // 是否为叶子节点
  isLeaf?: boolean;
  children?: TreeTransferDataType[];
  // 该节点是否为全选状态
  isCheckAll?: boolean;
  extraInfo?: { [key: string]: string };
  // 是否有更多更请求的数据。加载失败时设置其值为 error 显示重试按钮
  hasMore?: boolean | 'error';
  // 由外部指定该项是否选中，懒加载时可自定义设置
  checked?: boolean;
  icon?: React.ReactNode;
  nodeCheckAllExtra?: string | React.ReactNode;
  isClickCheckAll?: boolean;
  childrenData?: TreeTransferDataType[];
  [key: string]: unknown;
}